azure-policy-assistant/backend/requirements.txt
azure-functions
azure-cosmos
azure-storage-blob
openai
requests